Export Data (Example)

This data set resembles the export of an evaluation. Any data that is re-imported to append new data will need to follow this structure.

Usage

data(df_output_exportable_example)

Format

A data frame with 7 rows and 9 variables:

first_name

Speaker's first name.

last_name

Speaker's last name.

date_of_birth

The speaker's date of birth.

language_spoken

The language spoken during the evaluation.

gender

The speaker's gender.

date_of_evaluation

The date of the evaluation.

referent

Name of the referent associated with the response.

conversing

A binary indicator showing whether a speaker provided a thematic response to an object (1 = Yes, 0 = No).

labeling

A binary indicator showing whether the speaker could label an object (1 = Yes, 0 = No).

echoing

A binary indicator showing whether the speaker could repeat an object name (1 = Yes, 0 = No).

requesting

A binary indicator showing whether the speaker could request an object name (1 = Yes, 0 = No).