Prompt Hierarchy Table

This function generates a `data.frame` representing the prompt hierarchy, which outlines the structured intervention steps that a speaker is expected to follow during VOX Analysis. The hierarchy helps assess the progression of responses, providing insights into how prompts are followed or adjusted based on the speaker's responses. This table accepts only one unique evaluation date at a time.

Usage

table_prompt_hierarchy(df_summarized_response)

Arguments

df_summarized_response

A `data.frame` summarizing each verbal operant (Conversing, Labeling, Echoing, Requesting) for a given evaluation date. This summary provides a condensed view of responses and can be produced using the `util_summarize_response` function.

Examples

# Load example data
library(dplyr)
data("df_summarized_response_example")

# Filter to a single evaluation date, as the table works with one date at a time
dat <- df_summarized_response_example %>%
    filter(date_of_evaluation == max(date_of_evaluation))

# Generate the prompt hierarchy table
table_prompt_hierarchy(df_summarized_response = dat)
#> # A tibble: 14 × 5
#>    Percent Conversing Labeling Echoing Requesting
#>      <dbl> <chr>      <chr>    <chr>   <chr>     
#>  1   100   CELR       LERC     ELRC    RELC      
#>  2    91.7 -          LER      ELR     REL       
#>  3    83.3 CEL        LEC      ELC     -         
#>  4    75   -          LE       EL      -         
#>  5    66.7 CER        -        ERC     REC       
#>  6    58.4 -          -        ER      RE        
#>  7    58.3 CLR        LRC      -       RLC       
#>  8    50   CE         LR       EC      RL        
#>  9    41.7 -          -        E       -         
#> 10    41.6 CL         LC       -       -         
#> 11    33.3 -          L        -       -         
#> 12    25   CR         -        -       RC        
#> 13    16.7 -          -        -       R         
#> 14     8.3 C          -        -       -