Custom Numeric Input for Borderless Data Entry

This function creates a custom numeric input field, similar to `shiny::textInput`, but with a borderless design and no up / down arrows. It is used in `tag_table_response_inputs` to facilitate smooth, minimalistic numeric input in the VOX Analysis language assessment app.

Usage

tag_custom_input(
  input_id,
  value = 0,
  align_center = TRUE,
  placeholder = "Enter 1 or 0"
)

Arguments

input_id

A unique identifier for the input element.

value

A numeric value to be displayed as the default input.

align_center

Logical. If `TRUE`, centers the input alignment. Defaults to `FALSE`.

placeholder

A character string to display as placeholder text when the input is empty.

Value

A custom, borderless numeric input UI element without numeric increment arrows.