Render Word Document Report

This function generates a Word document report using the `printable_report.qmd` template found in the `inst` folder of the package. The report provides a formatted summary of VOX Analysis data, making it suitable for sharing and documentation purposes.

Usage

util_generate_word_doc_report(
  df_input_speaker_info = NULL,
  df_input_response,
  file_name
)

Arguments

df_input_speaker_info

A `data.frame` containing general information about a speaker (the subject of the VOX Analysis), as provided by the user. For an example, run `data("df_input_speaker_info")`.

df_input_response

A `data.frame` containing referent names and their associated responses, typically recorded by the listener. For example data, run `data("df_input_response_example")` or `data("df_input_response_previous_example")`.

file_name

A character string specifying the name of the exported file, including the `.docx` extension.

See also

Examples

if (FALSE) { # \dontrun{
# Load example data for report generation
data("df_input_speaker_info_example")
data("df_input_response_example")

# Generate a Word document report
util_generate_word_doc_report(
  df_input_speaker_info = df_input_speaker_info_example,
  df_input_response = df_input_response_example,
  file_name = "text.docx"
)
} # }