Summarize Responses by Type

This utility function aggregates responses by type (e.g., Conversing, Labeling, Echoing, Requesting) within the VOX Analysis application. It calculates the total count for each response type, providing a summary that can be used for further analysis or visualization.

Usage

util_summarize_response(df_input_response)

Arguments

df_input_response

A `data.frame` containing referent names and their associated responses, typically recorded by the listener. For example data, run `data("df_input_response_example")` or `data("df_input_response_previous_example")`.

Examples

# Load example data
data("df_input_response_example")

# Summarize responses by type
util_summarize_response(
  df_input_response = df_input_response_example
)
#> # A tibble: 1 × 5
#>   date_of_evaluation labeling echoing requesting conversing
#>   <date>                <dbl>   <dbl>      <dbl>      <dbl>
#> 1 2024-08-23                4       5          2          1